refactor(mobile): move navigation into separate files
