+++ mobile/navigation/RootNavigator/index.tsx
@@ -1,0 +1,100 @@
+import { FontAwesome } from '@expo/vector-icons';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import React from 'react';
+import { Pressable } from 'react-native';
+import useColorScheme from '../../hooks/useColorScheme';
+import { Colors } from 'react-native/Libraries/NewAppScreen';
+import ModalScreen from '../../screens/ModalScreen';
+import NotFoundScreen from '../../screens/NotFoundScreen';
+import TabOneScreen from '../../screens/TabOneScreen';
+import TabTwoScreen from '../../screens/TabTwoScreen';
+import {
+  RootStackParamList,
+  RootTabParamList,
+  RootTabScreenProps
+} from './types';
+
+const BottomTab = createBottomTabNavigator<RootTabParamList>();
+
+const BottomTabNavigator = () => {
+  const colorScheme = useColorScheme();
+
+  return (
+    <BottomTab.Navigator
+      initialRouteName="TabOne"
+      screenOptions={{
+        tabBarActiveTintColor: Colors[colorScheme].tint
+      }}
+    >
+      <BottomTab.Screen
+        name="TabOne"
+        component={TabOneScreen}
+        options={({ navigation }: RootTabScreenProps<'TabOne'>) => ({
+          title: 'Tab One',
+          tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
+          headerRight: () => (
+            <Pressable
+              onPress={() => navigation.navigate('Modal')}
+              style={({ pressed }) => ({
+                opacity: pressed ? 0.5 : 1
+              })}
+            >
+              <FontAwesome
+                name="info-circle"
+                size={25}
+                color={Colors[colorScheme].text}
+                style={{ marginRight: 15 }}
+              />
+            </Pressable>
+          )
+        })}
+      />
+      <BottomTab.Screen
+        name="TabTwo"
+        component={TabTwoScreen}
+        options={{
+          title: 'Tab Two',
+          tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />
+        }}
+      />
+      <BottomTab.Screen
+        name="Mobile"
+        component={TabTwoScreen}
+        options={{
+          title: 'Tab Two',
+          tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />
+        }}
+      />
+    </BottomTab.Navigator>
+  );
+};
+
+const TabBarIcon = (props: {
+  name: React.ComponentProps<typeof FontAwesome>['name'];
+  color: string;
+}) => {
+  return <FontAwesome size={30} style={{ marginBottom: -3 }} {...props} />;
+};
+
+export const RootNavigator: React.FC = () => {
+  const { Navigator, Group, Screen } =
+    createNativeStackNavigator<RootStackParamList>();
+  return (
+    <Navigator>
+      <Screen
+        name="Root"
+        component={BottomTabNavigator}
+        options={{ headerShown: false }}
+      />
+      <Screen
+        name="NotFound"
+        component={NotFoundScreen}
+        options={{ title: 'Oops!' }}
+      />
+      <Group screenOptions={{ presentation: 'modal' }}>
+        <Screen name="Modal" component={ModalScreen} />
+      </Group>
+    </Navigator>
+  );
+};
